Set a default matrix to use and allow overriding
The default matrix lists all of the options for the thing that is being
built. The default os and dist match what Travis-CI does, with os set to
"linux" and dist set to "precise".

The default matrix also sets the keys to search for in the .travis.yml
file to expand the build matrix. Each language has different keys about
what versions to use. Many use only one, the name of the language, but
others have a differently named key or have multiple keys. Support the
common, simple case by allowing a language to point to a list or tuple
to list the versions to run by default.