Decrease maximum step name to truncate long step names more
When the step name is too long then it gets truncated. However, when the
ellipses are appended the name is at the maximum length. If multiple
steps have the same name after the truncation, then the subsequent
mangling to add a "_N" pushes the step name over the maximum length.
Decrease the maximum step name by three to give room for "_NN".